Keep Chunk as an interface
