+++ common/src/main/java/us/myles/ViaVersion/api/minecraft/chunks/Chunk.java
@@ -1,23 +1,23 @@
 package us.myles.ViaVersion.api.minecraft.chunks;
 
 import com.github.steveice10.opennbt.tag.builtin.CompoundTag;
-import lombok.AllArgsConstructor;
-import lombok.Data;
 
 import java.util.List;
 
-@Data
-@AllArgsConstructor
-public class Chunk {
-    protected int x;
-    protected int z;
-    protected boolean groundUp;
-    protected int bitmask;
-    protected ChunkSection[] sections;
-    protected byte[] biomeData;
-    protected List<CompoundTag> blockEntities;
+public interface Chunk {
+    int getX();
 
-    public boolean isBiomeData() {
-        return biomeData != null;
-    }
+    int getZ();
+
+    boolean isBiomeData();
+
+    int getBitmask();
+
+    ChunkSection[] getSections();
+
+    byte[] getBiomeData();
+
+    List<CompoundTag> getBlockEntities();
+
+    boolean isGroundUp();
 }
