+++ common/src/main/java/us/myles/ViaVersion/api/minecraft/chunks/Chunk1_8.java
@@ -1,18 +1,16 @@
-package us.myles.ViaVersion.protocols.protocol1_9to1_8.chunks;
+package us.myles.ViaVersion.api.minecraft.chunks;
 
 import com.github.steveice10.opennbt.tag.builtin.CompoundTag;
 import lombok.Getter;
-import us.myles.ViaVersion.api.minecraft.chunks.Chunk;
-import us.myles.ViaVersion.api.minecraft.chunks.ChunkSection;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class Chunk1_9to1_8 extends Chunk {
+public class Chunk1_8 extends BaseChunk {
     @Getter
     private boolean unloadPacket = false;
 
-    public Chunk1_9to1_8(int x, int z, boolean groundUp, int bitmask, ChunkSection[] sections, byte[] biomeData, List<CompoundTag> blockEntities) {
+    public Chunk1_8(int x, int z, boolean groundUp, int bitmask, ChunkSection[] sections, byte[] biomeData, List<CompoundTag> blockEntities) {
         super(x, z, groundUp, bitmask, sections, biomeData, blockEntities);
     }
 
@@ -22,7 +20,7 @@ public class Chunk1_9to1_8 extends Chunk {
      * @param x coord
      * @param z coord
      */
-    public Chunk1_9to1_8(int x, int z) {
+    public Chunk1_8(int x, int z) {
         this(x, z, true, 0, new ChunkSection[16], null, new ArrayList<CompoundTag>());
         this.unloadPacket = true;
     }
